Extract server-artifact S3 handling into its own module
